* fixed a *nix build error
git-svn-id: https://ascentemubackup.svn.sourceforge.net/svnroot/ascentemubackup/trunk@1694 8ad44209-8f18-4ceb-bdfa-99d0be514da2

+++ src/game/Guild.cpp
@@ -168,7 +168,7 @@ void Guild::FillGuildRosterData(WorldPacket *data)
 	std::list<RankInfo*>::iterator ritr;
 
 	data->Initialize(SMSG_GUILD_ROSTER);
-	*data << GetGuildMembersCount();
+	*data << uint32(GetGuildMembersCount());
 	*data << m_motd;
 	*data << m_guildInfo;
 	*data << (uint32)m_rankList.size();
